+++ src/index.js
@@ -54,6 +54,7 @@ LayerUI.mixins = {
   ListSelection: require('./mixins/list-selection'),
   ListItemSelection: require('./mixins/list-item-selection'),
   FocusOnKeydown: require('./mixins/focus-on-keydown'),
+  AfterTextHandler: require('./mixins/after-text-handler'),
 };
 
 // If we don't expose global.layerUI then custom templates can not load and call window.layerUI.registerTemplate()